==> behavioural_analysis/2.preprocess.Rmd <==
---
title: "preprocess"
author: "RB"
date: "2023-09-25"
output: html_document
---
### Script to preprocess the behavioural data from the 'Two Currency Task' ###
# Adds variables that were not recorded during data collection

## load packages
```{r}
library(tidyverse)
```

# which data sets are we preprocessing (y=1/n=0)
```{r}
online = 1 # behavioural data from scan or outside scanner?
simulations = 1 # preprocess simulated data?
modelling = 1 # preprocess the modelling variables?
scan_data = 1 # preprocess the fMRI ROI outputs?
patient = 1 # preprocess patient data?
```

load in the data 
```{r}
if (online == 1){
    # list of subject numbers in strings
    nr_s <- c('02','03','04','05','06','07','08','09','10','11','12','13','14','15','16','17','18','19','20','21','22')
    
    # read in first data file
    a <- read.csv("raw_data/main/sub_02/raw1_02.csv") %>%
      select(!contains("fives")) 
    a['subject_nr'] <- nr_s[1]
    b <- read.csv("raw_data/main/sub_02/raw2_02.csv") %>%
      select(!contains("fives"))
    b['subject_nr'] <- nr_s[1]
    
    data <- rbind(a,b)
    
    # loop through the remaining files and bind to data 
    for (i in 2:length(nr_s)){
    file_1 <- str_replace_all("raw_data/main/sub_02/raw1_02.csv",'02',nr_s[i])
    file_2 <- str_replace_all("raw_data/main/sub_02/raw2_02.csv",'02',nr_s[i])
    
    a <- read.csv(file_1) %>%
      select(!contains("fives"))
    a['subject_nr'] <- nr_s[i]
    b <- read.csv(file_2) %>%
      select(!contains("fives"))
    b['subject_nr'] <- nr_s[i]
    
    data <- rbind(data,a,b)
    
    # rename as real data
    real_data <- data %>%
      rowid_to_column()

    }
    
} else if (online == 0) {
 # list the subject numbers in strings
  nr_s <- c('07','08','09','10','11','12','13','14','15','16','17','18','19','20','21','22',
            '23','24','25','26','27','28','29','30','31','32', '33', '34', '35', '36', '37', '38', '39', '40')
  
  # read in the first data file
  data <- read.csv("raw_data/behavioural/DA_07.csv")
  data['subject_nr'] <- nr_s[1]
  
  # loop through the remaining files and bind the data
  for (i in 2:length(nr_s)){
    file <- str_replace_all("raw_data/behavioural/DA_03.csv",'03',nr_s[i])  
    
    a <- read.csv(file)
    a['subject_nr'] <- nr_s[i]
    
    # check if 'bank_1' column exists, if it does that means bug occurred that split banks
    # need to remove bank_2
    # and rename bank_1 as bank
    x = "bank_1" %in% colnames(a)
    if (x == TRUE){
      a <- a %>%
        rename(bank=bank_1) %>%
        select(!bank_2)
    }
  data <- rbind(data,a)
  
  real_data <- data %>%
    rowid_to_column()
  }
}

```

Load simulated data
```{r}
if (simulations == 1){
  if (online == 1){
    # load simulation
    sim_data <- read.csv('raw_data/simulations/simulated_choice_data_M.csv') %>%
      rowid_to_column()
    sim_data$subject_nr <- rep(nr_s, each=500)
  } else if (online == 0){
    sim_data <- read.csv('raw_data/simulations/simulated_choice_data_FM_determ.csv') %>%
      rowid_to_column()
    sim_data$subject_nr <- rep(nr_s[1:21], each=1024)
  }
}
```

load patient data
```{r}
if (patient == 1){
  if (online == 0){
    p_data <- read.csv('raw_data/patient_AM_data/pt_01.csv') %>%
      rowid_to_column()
    p_data['subject_nr'] <- 41
  } else {
    a <- read.csv('raw_data/patient_AM_data/ps1_01.csv') %>%
        select(!contains("fives")) 
    b <- read.csv('raw_data/patient_AM_data/ps2_01.csv') %>%
        select(!contains("fives")) 
    c <- read.csv('raw_data/patient_AM_data/ps3_01.csv') %>%
        select(!contains("fives")) 
    p_data <- rbind(a,b,c) %>%
      rowid_to_column()
    p_data['subject_nr'] <- 23
  }
}
```

create the variables we forgot to add during the task
```{r}
# we want to run this block for both simulated and real data, so loop through these
datasets <- c('real','sim','p')

for (j in 1:length(datasets)){
  # first set data to real then to simulated
  if (datasets[j] == 'real'){
    data <- real_data
  } else if (datasets[j] == 'sim'){
    data <- sim_data
  } else if (datasets[j] == 'p'){
    data <- p_data
  }

  ## set empty statename variable and correct response 
  ## check if online or offline, offline has 1 extra state
  
  data['stateName'] <- 'none'
  data['correctDice'] <- 1 
  data['correct'] <- 0
  
  ## loop through data to give the states the corresponding state name
  # in future add this variable during the task
  if (online == 1){ 
    for (i in data$rowid){
      if (data[i,]$state == 3 | data[i,]$state == 8 | data[i,]$state == 12 | data[i,]$state == 18){
        data[i,]$stateName = 'fill'
      } else if (data[i,]$state == 4 | data[i,]$state == 9 | data[i,]$state == 13 | data[i,]$state == 19){
        data[i,]$stateName = 'extinction'
      } else if (data[i,]$state == 5 | data[i,]$state == 10 | data[i,]$state == 14 | data[i,]$state == 20){
        data[i,]$stateName = 'reinstatement'
      } else if (data[i,]$state == 6 | data[i,]$state == 11 | data[i,]$state == 16 | data[i,]$state == 21){
        data[i,]$stateName = 'post'
      } else if (data[i,]$state == 1 | data[i,]$state == 2 | data[i,]$state == 7 | data[i,]$state == 17){
        data[i,]$stateName = 'learning'
      } else if (data[i,]$state == 15){
        data[i,]$stateName = 'switch'
      }
    }
  } else if (online == 0){
      for (i in data$rowid){
        if (data[i,]$state == 4 | data[i,]$state == 9 | data[i,]$state == 13 | data[i,]$state == 19){
          data[i,]$stateName = 'fill'
        } else if (data[i,]$state == 5 | data[i,]$state == 10 | data[i,]$state == 14 | data[i,]$state == 20){
          data[i,]$stateName = 'extinction'
        } else if (data[i,]$state == 6 | data[i,]$state == 11 | data[i,]$state == 15 | data[i,]$state == 21){
          data[i,]$stateName = 'reinstatement'
        } else if (data[i,]$state == 7 | data[i,]$state == 12 | data[i,]$state == 17 | data[i,]$state == 22){
          data[i,]$stateName = 'post'
        } else if (data[i,]$state == 1 | data[i,]$state == 2 | data[i,]$state == 3 | data[i,]$state == 8 |
          data[i,]$state == 18){
          data[i,]$stateName = 'learning'
        } else if (data[i,]$state == 16){
          data[i,]$stateName = 'switch'
        }    
      }
  }
  
  ## now lets set the correct dice variable
  # this is the dice with the higher reward probability, if equal set to 3
  
  if (online == 1){
    for (i in data$rowid){
      if (data[i,]$state == 1 | data[i,]$state == 4 |  data[i,]$state == 5 | data[i,]$state == 7 | 
          data[i,]$state == 8 | data[i,]$state == 13 |data[i,]$state == 14 | data[i,]$state == 19 | 
          data[i,]$state == 20){
      data[i,]$correctDice = 2
      } else if (data[i,]$state == 6 |data[i,]$state == 11 | data[i,]$state == 16 | data[i,]$state == 21){
        data[i,]$correctDice = 3 # either response correct/equal reward probs
      }
    }
  } else if (online == 0){
      for (i in data$rowid){
        if (data[i,]$state == 2 | data[i,]$state == 5 |  data[i,]$state == 6 | data[i,]$state == 8 | 
            data[i,]$state == 9 | data[i,]$state == 14 |data[i,]$state == 15 | data[i,]$state == 20 | 
            data[i,]$state == 21){
        data[i,]$correctDice = 2
        } else if (data[i,]$state == 7 |data[i,]$state == 12 | data[i,]$state == 17 | data[i,]$state == 22){
          data[i,]$correctDice = 3 # either response correct/equal reward probs
        }
      }
  }
  
  ## now use this to show whether they chose this dice or not
  data$correct <- ifelse(data$correctDice == data$chosenDice, 1, 0) 
  
  if (datasets[j] == 'real'){
    real_data <- data
  } else if (datasets[j] == 'sim'){
    sim_data <- data
  } else if (datasets[j] == 'p'){
    p_data <- data
  }
}

```

```{r}
# set the real data to the data for the standard analysis
# we will bind the simulated data before the learning curve
if (online == 1)
  {
  # first load in the modelled data for chosen dev, using this later
  chosen_dev <- read.csv('modelling/summaries/m5_chosen_diff.csv')
  chosen_dev = chosen_dev$mean[501:11000]
  real_data <- cbind(real_data,chosen_dev)
  
  real_data <- real_data %>% 
    filter(!chosenDice == 'Inf') # filter time outs
  
  p_data <- p_data %>%
    filter(!chosenDice == 'Inf')
}

```

Save the clean data
```{r}
if (online == 1){
  write.csv(real_data, 'raw_data/clean/real_data_online.csv')
  write.csv(sim_data, 'raw_data/clean/sim_data_online.csv')
  write.csv(p_data, 'raw_data/clean/p_data_online.csv')
  } else {
  write.csv(real_data, 'raw_data/clean/real_data_offline.csv')
  write.csv(sim_data, 'raw_data/clean/sim_data_offline.csv')
  write.csv(p_data, 'raw_data/clean/p_data_offline.csv')
  }
```

### modelled values
```{r}
if (modelling == 1){
  # read in the model summary with the parameter estimates
  if (online == 1){
    params <- read.csv('modelling/summaries/full_model_patient.csv')
  } else if (online == 0){
    params <- read.csv('modelling/summaries/F0_model_summary.csv')
  }
  
  # learning rate
  lr <- params %>%
    select(X,mean) %>%
    filter(grepl('lr',X)) %>%
    filter(!X=='a_lr') %>% # remove hyperparameters of beta distrib.
    filter(!X=='b_lr') %>%
    rename(lr=mean)
  
  m <- params %>%
    select(X,mean) %>%
    filter(grepl('m',X)) %>%
    filter(!X=='a_m') %>% # remove hyperparameters of beta distrib.
    filter(!X=='b_m') %>%
    rename(m=mean)
  
  # inverse temperature
  tau <- params %>%
    select(X,mean) %>%
    filter(grepl('inv_temp',X)) %>%
    filter(!X=='k_tau') %>% # remove hyperparameters of gamma distrib.
    filter(!X=='theta_tau') %>%
    rename(inv_temp=mean)
  
  # remove sub 1 as not used in analysis
  if (online==1){
    lr <- lr[2:length(lr$lr),]
    tau <- tau[2:length(tau$inv_temp),]
    m <- m[2:length(m$m),]
  } 
  
  m <- m %>%
    mutate(mLog=log(m)) %>%
    mutate(mCent=(m-mean(m))) %>%
    mutate(mNorm=(m - min(m)) / (max(m) - min(m)))
  
  ## now bind them and save
  
  parameters <- as.data.frame(cbind(lr[,2],tau[,2],m[,2])) %>%
    rename(lr=V1, tau=V2, m=V3)
  
  if (online == 1)
  {
    write.csv(parameters, 'raw_data/clean/parameters_online.csv')
  } else {
    write.csv(parameters, 'raw_data/clean/parameters_offline.csv')
  }
}

```

### fmri values
```{r}
if (scan_data == 1){
  vmpfc_cd <- read.csv('C:/Users/Rowan/Desktop/oxford/analyse_timeCourses/vmPFC/chosen_diff_vmPFC.csv', header=FALSE, col.names = 'vmpfc_cd')
  vmpfc_m <- read.csv('C:/Users/Rowan/Desktop/oxford/analyse_timeCourses/vmPFC/motivation_vmPFC.csv', header=FALSE, col.names = 'vmpfc_m')
  vmpfc_rd <- read.csv('C:/Users/Rowan/Desktop/oxford/analyse_timeCourses/vmPFC/reward_diff_vmPFC.csv', header=FALSE, col.names = 'vmpfc_rd')
  vs_rd <- read.csv('C:/Users/Rowan/Desktop/oxford/analyse_timeCourses/VS/reward_diff_VS.csv', header=FALSE, col.names = 'vs_rd')
  vs_cd <- read.csv('C:/Users/Rowan/Desktop/oxford/analyse_timeCourses/VS/chosen_diff_VS.csv', header=FALSE, col.names = 'vs_cd')
  dmpfc_cd <- read.csv('C:/Users/Rowan/Desktop/oxford/analyse_timeCourses/dmPFC/chosen_diff_dmPFC.csv', header=FALSE, col.names = 'dmpfc_cd')
  lfp_cd <- read.csv('C:/Users/Rowan/Desktop/oxford/analyse_timeCourses/lFP/chosen_diff_lFP.csv', header=FALSE, col.names = 'lfp_cd')
  lvs_roi_rwd <- read.csv('C:/Users/Rowan/Desktop/oxford/analyse_timeCourses/left_vs_roi/reward_left_vs_roi.csv', header=FALSE, col.names = 'left_vs_roi_rwd')
  lvs_roi_rwd_d <- read.csv('C:/Users/Rowan/Desktop/oxford/analyse_timeCourses/left_vs_roi/reward_diff_left_vs_roi.csv', header=FALSE, col.names = 'left_vs_roi_rwd_d')
  
  fmri_vals <- as.data.frame(cbind(vmpfc_cd,vmpfc_m,vmpfc_rd,vs_cd,vs_rd,dmpfc_cd,lfp_cd,lvs_roi_rwd,lvs_roi_rwd_d))
  
  write.csv(fmri_vals, 'raw_data/clean/fmri_vals.csv')
}
```
